[Feat] 씬전환 작업 기능 구현중

+++ Assets/03.PrivateFolder/JangDJ/JDJ_Scripts/NoteSpawnPosController.cs
@@ -42,7 +42,7 @@ public class NoteSpawnPosController : MonoBehaviour
     private void SetSpawnPos()
     {
         float minOffsetX = 20f;
-        int speed = DataManager.Instance.GameSpeed;
+        float speed = DataManager.Instance.GameSpeed;
 
         for (int i = 1; i < 30; i++)
         {
